refactor: extract chat route handler from smart-server
